Ignore catalog (DB) name in multi-db handling logic for temp tables
The schema name is set to NULL for temp tables because schema name is silently ignored in SQL server. But we didn't set the catalog (DB) name to NULL when one is provided. This causes a crash when  we  use the schema name in a string comparison. This commit addresses this issue by also setting the catalog (DB) name to NULL for temp tables

Task: BABEL-2692 
Author: Huansong Fu <huansonf@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>