BABEL: Fix table type dependency on owner
Previously, when creating a table type, the underlying table had a dependency on both the table type and the the role that created it. The DROP OWNED BY <role> command will try to drop the table, and fail because the table depends on the table type. This commit removes the underlying table's dependency on the role, and instead adds a dependency for the table type on the role. This way, DROP OWNED BY <role> will try to drop the table type instead, and succeed in dropping both the table type and the underlying table.

Task: BABEL-2706
Author: Zitao Quan <qztao@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>
(cherry picked from commit 67f3ffb603b2965c5351b1c5b17c14ab5707c2c4)