Crash in exec_stmt_iterative due to trigger with error handling
Root cause: The problem was due to composite trigger with error
handling. When we run the DML, we record trigger events separately for
T-SQL triggers.

AfterTriggersTableData is recorded in composite triggers but a reference
is also maintained by PostgreSQL triggers. When we cleanup PostgreSQL
trigger data (due to rollback to savepoint etc.), we skip cleanup of
AfterTriggersTableData if composite triggers are active. Later,
composite triggers delete the AfterTriggersTableData.

However, if there is an error during DML execution, we first remove
composite triggers (which deletes AfterTriggersTableData) and later do
rollback to savepoint. In this case, the PostgreSQL trigger will have a
stale reference to AfterTriggersTableData since it is already deleted by
composite trigger and will lead to segment fault issues.

Solution:
In the event of error, EndCompositeTriggers() is being invoked before
AfterTriggerEndSubXact() which will cleanup
AfterTriggersTableData. But reference to the same is maintained by PG
trigger which in turn would have a stale reference. So, setting
reference to AfterTriggersTableData to the NIL explicitly if query_depth
of
PG trigger is more than query_depth of the composite trigger.

Task: BABEL-2845
Author: Dipesh Dhameliya dddhamel@amazon.com
Signed-off-by: Dipesh Dhameliya dddhamel@amazon.com