BABEL-3193, BABEL-3194 and BABEL-3208: BCP Related Bug-fixes
In this commit we fix 3 issues:

1. Bcp does not handle ‘db..table’ syntax. The issue was pertaining to
sp_describe_first_resultset where SPI API was not able to resolve the
DB..TABLE name. To fix this we are making use of ANTLR Parser to resolve
the full name, in addition it also provides syntax checks.

2. Bcp-in does not handle identity column. The issue was with views created
from select statement does not inherit the constraints from the base columns
which was the reason why sp_describe_first_resultset sent wrong values for
identity columns. To Fix this we lookup the catalog at the time of view
creation in TSQL Dialect to initialise the constraints for it.

3. We had restricted sp_describe_first_result_set to only work with SELECT
statements and for non-select statements we threw an error. With this commit
we return 0 rows for any valid non-select query.

Also added JDBC tests