Fix MVU failure when a view contains sys.suser_sid()
If a view from v13.6 contains sys.suser_sid() and its
parameter is not explicitly casted to sys.sysname, then Major Version
Upgrade (MVU) to v14.latest fails. This is because we had assigned typmod for
all domains in T-SQL dialect in v13.x, which was fixed in v14.latest.

To prevent the MVU failure, we skip an error related to sys.sysname's
typmod when babelfishpg_tsql.dump_restore GUC is ON.

Task: BABEL-3402
Signed-off-by: Jungkook Lee <jungkook@amazon.com>