BABEL: Fix Major-version-upgrade issue when column name has comma (#37)
* Fix Major-version-upgrade issue when column name has comma

If comma is included in column name, bbf_original_name attoptions is
parsed wrongly and it cause a failure of MVU. Before the fix,
attoptions is collpased into a string so it is hard to split into array
elements if comma is involved. This fix will access pg_attribute to get
attoption text[] so it will not cause a parse-issue.

Task: BABEL-3121
Authored-by: Sangil Song <sonsangi@amazon.com>
Signed-off-by: Sangil Song <sonsangi@amazon.com>

* Refactoring

Signed-off-by: Jungkook Lee <jungkook@amazon.com>

Co-authored-by: Sangil Song <sonsangi@amazon.com>
Co-authored-by: Jungkook Lee <jungkook@amazon.com>
(cherry picked from commit 72148d4af1fda9ebb6b7f27a685f004d039c3050)